all works generically
added port maps to each entity.
now, when a server wants to send an http request, it looks at the state
machine's map for the right router for to forward the message.
router receive a 'rout' request and find the right server host and port
at its port map as well.