digraph
json parser now parse the data from json files into a erlang digraph.
edges in the graph are the connections, vertices are the entities and
their value is their host and port number. this way any two entities in
the network can find the shortest path in the network to communicate
with each other.