iot
now this application can start on startup using the script startUp.sh,
see readme in /initiator
also now nerlnet waits for a message from python about the whereabouts
of the json file to be parsed.
also added an option for python to search in the subnet for available
devices. see nerknetInit.py - just a structure for the real deal