Fixed several bugs
statistics option 1: plt.show() was indented backwards, to create only a
single figure.

statistics option 2: SciPy does not like floating point labels, because
it sees them as "continuous" values, but expects classification and not
regression to create a Confusion Matrix.
Therefore, I casted the labels to str type, so the Confusion Matrix will
be created.