ci: add pipeline
This YAML file defines a workflow that runs on every  pull request targeting the master branch. The workflow consists of a single job called build, which runs on the latest version of Ubuntu.

The job contains three steps:

The first step checks out the repository code.
The second step installs the curl package, which may be needed by the NerlnetInstall.sh script.
The third step makes the NerlnetInstall.sh script executable and runs it.